WIP: better liveness
A thread can be associated to a non-live role, in which case it can take
silent steps until it finishes, but it cannot take observable steps. The
hope is that it lets us handle the ending of threads after it took their
last step in more usable manner than currently.